06.08.2022 Inheritance study.

+++ src/_08_Agu_OOP/Inheritance/School/StudentEntities.java
@@ -5,7 +5,7 @@ public class StudentEntities {
 
         public double calculateGradeAvg( int grade1,int grade2, int grade3 ){
 
-            return (double) (grade1+grade2+grade3)/3;
+            return (float) (grade1+grade2+grade3)/3;
 
         }
